emails/requirements.txt
sendgrid
flask
python-dotenv
